Move test entities generators to mocks
